+++ include/zylib_error.h
@@ -36,24 +36,24 @@ extern "C"
     __attribute__((nonnull)) void zylib_error_destruct(zylib_error_t **err);
     __attribute__((nonnull)) void zylib_error_clear(zylib_error_t *err);
     __attribute__((nonnull(1, 3, 5))) zylib_return_t zylib_error_push_first(zylib_error_t *err, int64_t code,
-                                                                            const char *file, size_t line,
+                                                                            const char *file, uint64_t line,
                                                                             const char *function,
                                                                             const zylib_box_t *box);
     __attribute__((nonnull(1, 3, 5))) zylib_return_t zylib_error_push_last(zylib_error_t *err, int64_t code,
-                                                                           const char *file, size_t line,
+                                                                           const char *file, uint64_t line,
                                                                            const char *function,
                                                                            const zylib_box_t *box);
     __attribute__((nonnull)) void zylib_error_discard_first(zylib_error_t *err);
     __attribute__((nonnull)) void zylib_error_discard_last(zylib_error_t *err);
     __attribute__((nonnull)) const zylib_error_box_t *zylib_error_peek_first(const zylib_error_t *err);
     __attribute__((nonnull)) const zylib_error_box_t *zylib_error_peek_last(const zylib_error_t *err);
-    __attribute__((nonnull)) size_t zylib_error_size(const zylib_error_t *err);
+    __attribute__((nonnull)) uint64_t zylib_error_size(const zylib_error_t *err);
     __attribute__((nonnull)) _Bool zylib_error_is_empty(const zylib_error_t *err);
     __attribute__((nonnull)) int64_t zylib_error_box_peek_code(const zylib_error_box_t *bx);
     __attribute__((nonnull)) const char *zylib_error_box_peek_file(const zylib_error_box_t *bx);
-    __attribute__((nonnull)) size_t zylib_error_box_peek_line_number(const zylib_error_box_t *bx);
+    __attribute__((nonnull)) uint64_t zylib_error_box_peek_line_number(const zylib_error_box_t *bx);
     __attribute__((nonnull)) const char *zylib_error_box_peek_function(const zylib_error_box_t *bx);
-    __attribute__((nonnull(1))) const void *zylib_error_box_peek_opaque(const zylib_error_box_t *bx, size_t *size);
+    __attribute__((nonnull(1))) const void *zylib_error_box_peek_opaque(const zylib_error_box_t *bx, uint64_t *size);
 
 #ifdef __cplusplus
 }
